+++ backend/app/schemas/service.py
@@ -43,7 +43,7 @@ class HealthCheckBase(BaseModel):
     body: Optional[str] = None
     expected_status_code: int
     timeout_seconds: int = 30
-    check_interval_minutes: int = 1
+    check_interval_minutes: float = 1
     json_path: Optional[str] = None
     expected_value: Optional[str] = None
     enabled: bool = True
@@ -62,7 +62,7 @@ class HealthCheckUpdate(BaseModel):
     body: Optional[str] = None
     expected_status_code: Optional[int] = None
     timeout_seconds: Optional[int] = None
-    check_interval_minutes: Optional[int] = None
+    check_interval_minutes: Optional[float] = None
     json_path: Optional[str] = None
     expected_value: Optional[str] = None
     enabled: Optional[bool] = None
@@ -97,7 +97,7 @@ class VersionCheckBase(BaseModel):
     url: str
     json_path: str
     timeout_seconds: int = 30
-    check_interval_minutes: int = 1
+    check_interval_minutes: float = 1
     enabled: bool = True
 
 
@@ -110,7 +110,7 @@ class VersionCheckUpdate(BaseModel):
     url: Optional[str] = None
     json_path: Optional[str] = None
     timeout_seconds: Optional[int] = None
-    check_interval_minutes: Optional[int] = None
+    check_interval_minutes: Optional[float] = None
     enabled: Optional[bool] = None
